Move hockey example out of the comprl package
Move the hockey example out of the comprl package and split into two
parts:
- comprl-hockey-game: Contains the game implementation and the
  config.toml
- comprl-hockey-agent: Contains the example agent as a fully-fledged
  Python package.  It's a bit overkill for this example on its own but
  will make it easier for users to extend if they base their own agent
  implementation on this package.

This change will make it easier to provide users with the example agent
package and also allows to remove the (rather expensive) dependency on
the `hockey` package from the core comprl.
